update slot and unit save actions
