+++ src/actions/newItemsActions.js
@@ -2,33 +2,10 @@ import {
 	UPDATE_NEW_ITEM,
 	RESET_NEW_ITEM,
 	RESET_ALL_NEW_ITEMS,
-	ADD_ITEM,
 } from 'constants/actionTypes'
-import Helper from 'helpers/miscHelpers'
-import { numStringCPMtoImpression } from 'helpers/numbers'
-import {
-	t,
-	selectNewItems,
-	selectAuthSig,
-	selectMainToken,
-	selectNewAdSlot,
-	selectNewAdUnit,
-} from 'selectors'
-import { Base, Models, AdSlot, AdUnit } from 'adex-models'
-import {
-	updateSpinner,
-	addToast,
-	handleAfterValidation,
-	getImgsIpfsFromBlob,
-} from 'actions'
-import {
-	uploadImage,
-	postAdUnit,
-	postAdSlot,
-	updateAdSlot,
-	updateAdUnit,
-	updateCampaign,
-} from 'services/adex-market/actions'
+import { selectNewItems } from 'selectors'
+import { Base, Models } from 'adex-models'
+import { updateSpinner, handleAfterValidation } from 'actions'
 
 export function updateNewItem(item, newValues, itemType, objModel) {
 	item = Base.updateObject({ item, newValues, objModel })
@@ -157,131 +134,5 @@ export function completeItem({
 		await updateSpinner(validateId, false)(dispatch)
 
 		await handleAfterValidation({ isValid, onValid, onInvalid })
-	}
-}
-
-export function addSlot() {
-	return async function(dispatch, getState) {
-		try {
-			const state = getState()
-			const item = selectNewAdSlot(state)
-			const newItem = { ...item }
-			const authSig = selectAuthSig(state)
-			const mainToken = selectMainToken()
-			let fallbackUnit = null
-			if (newItem.temp.useFallback) {
-				const imageIpfs = (await getImgsIpfsFromBlob({
-					tempUrl: newItem.temp.tempUrl,
-					authSig,
-				})).ipfs
-
-				const unit = new AdUnit({
-					type: newItem.type,
-					mediaUrl: `ipfs://${imageIpfs}`,
-					targetUrl: newItem.targetUrl,
-					mediaMime: newItem.temp.mime,
-					created: Date.now(),
-					title: newItem.title,
-					description: newItem.description,
-					targeting: [],
-					tags: [],
-					passback: true,
-				})
-
-				const resUnit = await postAdUnit({
-					unit: unit.marketAdd,
-					authSig,
-				})
-
-				fallbackUnit = resUnit.ipfs
-			}
-
-			newItem.fallbackUnit = fallbackUnit
-			newItem.created = Date.now()
-
-			if (newItem.minPerImpression) {
-				newItem.minPerImpression = {
-					[mainToken.address]: numStringCPMtoImpression({
-						numStr: newItem.minPerImpression,
-						decimals: mainToken.decimals,
-					}),
-				}
-			}
-
-			const resItem = await postAdSlot({
-				slot: new AdSlot(newItem).marketAdd,
-				authSig,
-			})
-
-			dispatch({
-				type: ADD_ITEM,
-				item: new AdSlot(resItem).plainObj(),
-				itemType: 'AdSlot',
-			})
-
-			addToast({
-				type: 'accept',
-				label: t('SUCCESS_CREATING_ITEM', { args: ['AdSlot', newItem.title] }),
-				timeout: 50000,
-			})(dispatch)
-		} catch (err) {
-			console.error('ERR_CREATING_ITEM', err)
-			addToast({
-				type: 'cancel',
-				label: t('ERR_CREATING_ITEM', {
-					args: ['AdSlot', Helper.getErrMsg(err)],
-				}),
-				timeout: 50000,
-			})(dispatch)
-			throw new Error('ERR_CREATING_ITEM', err)
-		}
-	}
-}
-
-// register item
-export function addUnit() {
-	return async function(dispatch, getState) {
-		try {
-			const state = getState()
-			const item = selectNewAdUnit(state)
-			const newItem = { ...item }
-			const authSig = selectAuthSig(state)
-			const imageIpfs = (await getImgsIpfsFromBlob({
-				tempUrl: newItem.temp.tempUrl,
-				authSig,
-			})).ipfs
-
-			newItem.mediaUrl = `ipfs://${imageIpfs}`
-			newItem.mediaMime = newItem.temp.mime
-			newItem.created = Date.now()
-
-			const resItem = await postAdUnit({
-				unit: new AdUnit(newItem).marketAdd,
-				authSig,
-			})
-
-			dispatch({
-				type: ADD_ITEM,
-				item: new AdUnit(resItem).plainObj(),
-				itemType: 'AdUnit',
-			})
-
-			addToast({
-				dispatch: dispatch,
-				type: 'accept',
-				toastStr: 'SUCCESS_CREATING_ITEM',
-				args: ['AdUnit', newItem.title],
-			})
-		} catch (err) {
-			console.error('ERR_CREATING_ITEM', err)
-			addToast({
-				type: 'cancel',
-				label: t('ERR_CREATING_ITEM', {
-					args: ['AdUnit', Helper.getErrMsg(err)],
-				}),
-				timeout: 50000,
-			})(dispatch)
-			throw new Error('ERR_CREATING_ITEM', err)
-		}
 	}
 }
