+++ src/actions/slotActions.js
@@ -5,13 +5,22 @@ import {
 	validateNumberString,
 	validateMediaSize,
 	updateNewSlot,
+	addToast,
 	validate,
+	getImgsIpfsFromBlob,
 } from 'actions'
 import { numStringCPMtoImpression } from 'helpers/numbers'
-import { selectMainToken, selectNewAdSlot, t } from 'selectors'
-import { schemas } from 'adex-models'
-import { verifyWebsite } from 'services/adex-market/actions'
+import { selectMainToken, selectNewAdSlot, selectAuthSig, t } from 'selectors'
+import { schemas, AdSlot, AdUnit } from 'adex-models'
+import {
+	verifyWebsite,
+	postAdUnit,
+	postAdSlot,
+} from 'services/adex-market/actions'
 import { getWidAndHightFromType } from 'helpers/itemsHelpers'
+
+import { ADD_ITEM } from 'constants/actionTypes'
+import Helper from 'helpers/miscHelpers'
 
 const { adSlotPost, adUnitPost } = schemas
 
@@ -171,5 +180,83 @@ export function validateNewSlotPassback({
 
 		await handleAfterValidation({ isValid, onValid, onInvalid })
 		await updateSpinner(validateId, false)(dispatch)
+	}
+}
+
+export function saveSlot() {
+	return async function(dispatch, getState) {
+		try {
+			const state = getState()
+			const item = selectNewAdSlot(state)
+			const newItem = { ...item }
+			const authSig = selectAuthSig(state)
+			const mainToken = selectMainToken()
+			let fallbackUnit = null
+			if (newItem.temp.useFallback) {
+				const imageIpfs = (await getImgsIpfsFromBlob({
+					tempUrl: newItem.temp.tempUrl,
+					authSig,
+				})).ipfs
+
+				const unit = new AdUnit({
+					type: newItem.type,
+					mediaUrl: `ipfs://${imageIpfs}`,
+					targetUrl: newItem.targetUrl,
+					mediaMime: newItem.temp.mime,
+					created: Date.now(),
+					title: newItem.title,
+					description: newItem.description,
+					targeting: [],
+					tags: [],
+					passback: true,
+				})
+
+				const resUnit = await postAdUnit({
+					unit: unit.marketAdd,
+					authSig,
+				})
+
+				fallbackUnit = resUnit.ipfs
+			}
+
+			newItem.fallbackUnit = fallbackUnit
+			newItem.created = Date.now()
+
+			if (newItem.minPerImpression) {
+				newItem.minPerImpression = {
+					[mainToken.address]: numStringCPMtoImpression({
+						numStr: newItem.minPerImpression,
+						decimals: mainToken.decimals,
+					}),
+				}
+			}
+
+			const resItem = await postAdSlot({
+				slot: new AdSlot(newItem).marketAdd,
+				authSig,
+			})
+
+			dispatch({
+				type: ADD_ITEM,
+				item: new AdSlot(resItem).plainObj(),
+				itemType: 'AdSlot',
+			})
+
+			addToast({
+				type: 'accept',
+				label: t('SUCCESS_CREATING_ITEM', { args: ['AdSlot', newItem.title] }),
+				timeout: 50000,
+			})(dispatch)
+		} catch (err) {
+			console.error('ERR_CREATING_ITEM', err)
+			addToast({
+				type: 'cancel',
+				label: t('ERR_CREATING_ITEM', {
+					args: ['AdSlot', Helper.getErrMsg(err)],
+				}),
+				timeout: 50000,
+			})(dispatch)
+			throw new Error('ERR_CREATING_ITEM', err)
+		}
 	}
 }
